+++ ngapp/projects/gramadoir-error-line-chart/package.json
@@ -2,8 +2,8 @@
   "name": "gramadoir-error-line-chart",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^13.3.0",
-    "@angular/core": "^13.3.0"
+    "@angular/common": "^13.3.4",
+    "@angular/core": "^13.3.4"
   },
   "dependencies": {
     "@angular/platform-browser": "^13.3.11",
