+++ ngapp/projects/gramadoir-error-line-chart/src/lib/gramadoir-error-line-chart.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { GramadoirErrorLineChartComponent } from './gramadoir-error-line-chart.component';
 import { NgChartsModule } from 'ng2-charts';
 
+
 @NgModule({
   declarations: [
     GramadoirErrorLineChartComponent
