+++ ngapp/src/app/grammar.service.ts
@@ -11,8 +11,7 @@ import { HighlightTag } from 'angular-text-input-highlight';
 import { Story } from 'app/story';
 import { EngagementService } from 'app/engagement.service';
 import { EventType } from 'app/event';
-import { VowelAgreementIndex } from './services/quill-highlight.service';
-import config from 'abairconfig';
+import { GramadoirRuleId } from '../gramadoir-rule-id';
 
 export type DisagreeingVowelIndices = {
   broadFirst: boolean;
@@ -172,95 +171,6 @@ export const ReadableGramadoirRuleIds = {
   default: { ga: 'Error type unkown', en: 'Error type unkown' },
 };
 
-const GRAMADOIR_RULE_ID_VALUES: string[] = [
-  'CAIGHDEAN',
-  'SEIMHIU',
-  'CLAOCHLU',
-  'CUPLA',				  // Unusual combination of words
-  'DROCHMHOIRF',		// Derived incorrectly from the root {arg 1}
-  'NEAMHCHOIT',     // = same as next
-  'NAMMENOWGH',     // = same as next
-  'UNCOMMON',       // 'Valid word but extremely
-                    // rare in actual usage.
-                    // Is this the word you want?'
-  'DUBAILTE',       // 'Repeated word'
-  'CUPLA',          // = same as next
-  'KESUNYANS',      // 'Unusual combination of words'
-  'IOLRA',          // 'The plural form is required here'
-  'UATHA',          // 'The singular form is required here'
-  'AIDIOLRA',       // 'Plural adjective required'
-  'BREISCHEIM',     // 'Comparative adjective required'
-  'NEEDART',        // 'Definite article required'
-  'BADART',         // 'Unnecessary use of the definite article'
-  'ONEART',         // 'No need for the first definite article'
-  'NOGENITIVE',     // 'Unnecessary use of the genitive case'
-  'GENITIVE',       // 'The genitive case is required here'
-  'PRESENT',        // 'You should use the present tense here'
-  'CONDITIONAL',    //  'You should use the conditional here'
-  'NOSUBJ',         //  'It seems unlikely that you intended to use the subjunctive here'
-
-  // # TRANSLATORS: You can use whatever kind of quotes you prefer for your locale
-  // # around the variable \\1.  You should keep the double backslash before the 1,
-  // # but there is no need to escape the quotes the way they are in the msgid.
-  'INPHRASE',       // 'Usually used in the set phrase \/\1\/'
-  'BACHOIR',
-  'EWNHEANS',       //  'You should use \/\1\/ here instead'
-  'CAIGHDEAN',      // 'Non-standard form of \/\1\/'
-  'CAIGHMOIRF',     // 'Derived from a non-standard form of \/\1\/'
-  'DROCHMHOIRF',    // 'Derived incorrectly from the root \/\1\/'
-  'ANAITHNID',
-  'ANKOTHVOS',
-  'UNKNOWN', // 'Unknown word'
-
-  // # TRANSLATORS: \\1 is substituted with one or more suggested replacements
-  'MOLADH',         //  'Unknown word: \/\1\/?'
-  'IONADAI',        //  'Valid word but \/\1\/ is more common'
-  'MOIRF',          //  'Not in database but apparently formed from the root \/\1\/'
-  'NIGA',           //  'The word \/\1\/ is not needed'
-  'MICHEART',       // 'Do you mean \/\1\/?'
-  'MIMHOIRF',       // 'Derived form of common misspelling \/\1\/?'
-  'COMHFHOCAL',     // 'Not in database but may be a compound \/\1\/?'
-  'COMHCHAIGH',     // 'Not in database but may be a non-standard compound \/\1\/?'
-  'GRAM',           // 'Possibly a foreign word (the sequence \/\1\/ is highly improbable)'
-  'GENDER',
-  'GENRE',   // 'Gender disagreement'
-  'NUMBER',
-  'NOMBRE',  // 'Number disagreement'
-  'CASE',           // 'Case disagreement'
-
-  // # TRANSLATORS: Rules specific to the Irish language (ga)
-  'PREFIXH',        // 'Prefix \/h\/ missing'
-  'PREFIXT',        // 'Prefix \/t\/ missing'
-  'PREFIXD',        // 'Prefix \/d\'\/ missing'
-  'NIAITCH',        // 'Unnecessary prefix \/h\/'
-  'NITEE',          // 'Unnecessary prefix \/t\/'
-  'NIDEE',          // 'Unnecessary prefix \/d\'\/'
-  'NIBEE',          // 'Unnecessary prefix \/b\'\/'
-
-  // # TRANSLATORS: "Mutation" refers to either "lenition" or "eclipsis" (see below)
-  'NICLAOCHLU',     // 'Unnecessary initial mutation'
-  'CLAOCHLU',       // 'Initial mutation missing'
-
-  // # TRANSLATORS: "Lenition" is the softening of an initial consonant in Irish.
-  // # It is indicated in writing by the addition of an "h": e.g. "bean" -> "bhean"
-  'NISEIMHIU',      // 'Unnecessary lenition'
-  'NIDARASEIMHIU',  // 'The second lenition is unnecessary'
-  'WEAKSEIMHIU',    // 'Often the preposition \/\1\/ causes lenition, but this case is unclear'
-  'SEIMHIU',        // 'Lenition missing'
-
-  // # TRANSLATORS: "Eclipsis" is, like lenition, a phonetic change applied to
-  // # initial consonants in Irish.  It is indicated in writing by the addition
-  // # of the eclipsing consonant as a prefix: e.g. "bean" -> "mbean"
-  'NIURU',          // 'Unnecessary eclipsis'
-  'URU',            // 'Eclipsis missing'
-  'NODATIVE',       // 'The dative is used only in special phrases'
-  'RELATIVE',       // 'The dependent form of the verb is required here'
-  'ABSOLUTE',       // 'Unnecessary use of the dependent form of the verb'
-  'SYNTHETIC',      // 'The synthetic (combined) form, ending in \/\1\/, is often used here'
-
-  'default']; // Add more valid types here
-
-
 const IGNORE_THESE_GRAMMAR_MESSAGES: string[] = [
   // # TRANSLATORS: Rules specific to the Cornish language (kw)
   'NESSATREYLYANS',     // 'Second (soft) mutation missing'
@@ -273,10 +183,6 @@ const IGNORE_THESE_GRAMMAR_MESSAGES: string[] = [
   // # TRANSLATORS: Rules specific to the Igbo language (ig)
   'VOWELHARMONY',       // 'This word violates the rules of Igbo vowel harmony'
 ];
-// Makes a union type using the values
-// from the array above, see:
-// https://www.typescriptlang.org/docs/handbook/2/indexeda
-export type GramadoirRuleId = typeof GRAMADOIR_RULE_ID_VALUES[number];
 
 export enum LANGUAGE {
   ENGLISH = 0,
